pyproject.toml
[project]
name = "quant-py"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "jkitzler", email = "0.chirp.sleeved@icloud.com" }
]
requires-python = ">=3.14"
dependencies = [
    "numpy>=2.3.5",
    "pendulum>=3.1.0",
    "python-dateutil>=2.9.0.post0",
    "types-python-dateutil>=2.9.0.20251115",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "coverage>=7.12.0",
    "pyrefly>=0.43.1",
    "pytest>=9.0.1",
    "pytest-cov>=7.0.0",
    "pytest-xdist>=3.8.0",
    "ruff>=0.14.7",
]
